+++ Assets/VRMShaders/Runtime/TextureImportName.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d005850c7abbb24582e9788657afd35
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
